+++ frontend/common/services/useGroup.ts
@@ -1,6 +1,8 @@
 import { Res } from 'common/types/responses'
 import { Req } from 'common/types/requests'
 import { service } from 'common/service'
+import { getStore } from 'common/store'
+import Utils from 'common/utils/utils'
 
 export const groupService = service
   .enhanceEndpoints({
@@ -8,6 +10,39 @@ export const groupService = service
   })
   .injectEndpoints({
     endpoints: (builder) => ({
+      createGroup: builder.mutation<Res['group'], Req['createGroup']>({
+        invalidatesTags: [{ id: 'LIST', type: 'Group' }],
+        queryFn: async (query, { dispatch }, _, baseQuery) => {
+          //Create the group
+          const { data, error } = await baseQuery({
+            body: query.data,
+            method: 'POST',
+            url: `organisations/${query.orgId}/groups/`,
+          })
+          if (error) {
+            return { error }
+          }
+          //Add the members
+          if (query.users?.length) {
+            const { error } = await baseQuery({
+              body: { user_ids: query.users.map((u) => u.id) },
+              method: 'POST',
+              url: `organisations/${query.orgId}/groups/${data.id}/`,
+            })
+          }
+          // Make the admins
+          await Promise.all(
+            (query.usersToAddAdmin || []).map((v) =>
+              createGroupAdmin(getStore(), {
+                group: data.id,
+                orgId: query.orgId,
+                user: v,
+              }),
+            ),
+          )
+          return { data }
+        },
+      }),
       createGroupAdmin: builder.mutation<
         Res['groupAdmin'],
         Req['createGroupAdmin']
@@ -50,9 +85,62 @@ export const groupService = service
       }),
       getGroups: builder.query<Res['groups'], Req['getGroups']>({
         providesTags: [{ id: 'LIST', type: 'Group' }],
-        query: (query) => ({
-          url: `organisations/${query.orgId}/groups/?page=${query.page}`,
+        query: ({ orgId, ...rest }) => ({
+          url: `organisations/${orgId}/groups/?${Utils.toParam({ ...rest })}`,
         }),
+      }),
+      updateGroup: builder.mutation<Res['group'], Req['updateGroup']>({
+        invalidatesTags: (res) => [
+          { id: 'LIST', type: 'Group' },
+          { id: res?.id, type: 'Group' },
+        ],
+        queryFn: async (query, { dispatch }, _, baseQuery) => {
+          //Create the group
+          const { data, error } = await baseQuery({
+            body: query.data,
+            method: 'PUT',
+            url: `organisations/${query.orgId}/groups/${query.data.id}`,
+          })
+          if (error) {
+            return { error }
+          }
+          //Add the members
+          if (query.users?.length) {
+            await baseQuery({
+              body: { user_ids: query.data.users.map((u) => u.id) },
+              method: 'POST',
+              url: `organisations/${query.orgId}/groups/${data.id}/add-users/`,
+            })
+          }
+          if (query.usersToRemove?.length) {
+            await baseQuery({
+              body: { user_ids: query.usersToRemove },
+              method: 'POST',
+              url: `organisations/${query.orgId}/groups/${data.id}/remove-users/`,
+            })
+          }
+          // Make the admins
+          await Promise.all(
+            (query.usersToAddAdmin || []).map((v) =>
+              createGroupAdmin(getStore(), {
+                group: data.id,
+                orgId: query.orgId,
+                user: v,
+              }),
+            ),
+          )
+
+          await Promise.all(
+            (query.usersToRemoveAdmin || []).map((v) =>
+              deleteGroupAdmin(getStore(), {
+                group: data.id,
+                orgId: query.orgId,
+                user: v,
+              }),
+            ),
+          )
+          return { data }
+        },
       }),
       // END OF ENDPOINTS
     }),
@@ -106,15 +194,34 @@ export async function getGroup(
 ) {
   return store.dispatch(groupService.endpoints.getGroup.initiate(data, options))
 }
+export async function createGroup(
+  store: any,
+  data: Req['createGroup'],
+  options?: Parameters<typeof groupService.endpoints.createGroup.initiate>[1],
+) {
+  return store.dispatch(
+    groupService.endpoints.createGroup.initiate(data, options),
+  )
+}
+export async function updateGroup(
+  store: any,
+  data: Req['updateGroup'],
+  options?: Parameters<typeof groupService.endpoints.updateGroup.initiate>[1],
+) {
+  return store.dispatch(
+    groupService.endpoints.updateGroup.initiate(data, options),
+  )
+}
 // END OF FUNCTION_EXPORTS
 
 export const {
   useCreateGroupAdminMutation,
+  useCreateGroupMutation,
   useDeleteGroupAdminMutation,
   useDeleteGroupMutation,
-
   useGetGroupQuery,
   useGetGroupsQuery,
+  useUpdateGroupMutation,
   // END OF EXPORTS
 } = groupService