+++ frontend/web/components/FeatureRow.js
@@ -15,6 +15,7 @@ import Tag from './tags/Tag'
 import Button from './base/forms/Button'
 import SegmentOverridesIcon from './SegmentOverridesIcon'
 import IdentityOverridesIcon from './IdentityOverridesIcon'
+import StaleFlagWarning from './StaleFlagWarning'
 
 export const width = [200, 70, 55, 70, 450]
 class TheComponent extends Component {
@@ -57,7 +58,7 @@ class TheComponent extends Component {
     const { feature, tab } = Utils.fromParam()
     const { id } = projectFlag
     if (`${id}` === feature) {
-      this.editFeature(projectFlag, environmentFlags[id], tab)
+      this.editFeature(projectFlag, environmentFlags[id])
     }
   }
 
@@ -88,7 +89,9 @@ class TheComponent extends Component {
     history.replaceState(
       {},
       null,
-      `${document.location.pathname}?feature=${projectFlag.id}`,
+      `${document.location.pathname}?feature=${projectFlag.id}&tab=${
+        tab || Utils.fromParam().tab || 'value'
+      }`,
     )
     openModal(
       <Row>
@@ -104,6 +107,7 @@ class TheComponent extends Component {
         </Button>
       </Row>,
       <CreateFlagModal
+        history={this.context.router.history}
         environmentId={this.props.environmentId}
         projectId={this.props.projectId}
         projectFlag={projectFlag}
@@ -200,14 +204,22 @@ class TheComponent extends Component {
               <SegmentOverridesIcon
                 onClick={(e) => {
                   e.stopPropagation()
-                  this.editFeature(projectFlag, environmentFlags[id], 1)
+                  this.editFeature(
+                    projectFlag,
+                    environmentFlags[id],
+                    'segment-overrides',
+                  )
                 }}
                 count={projectFlag.num_segment_overrides}
               />
               <IdentityOverridesIcon
                 onClick={(e) => {
                   e.stopPropagation()
-                  this.editFeature(projectFlag, environmentFlags[id], 1)
+                  this.editFeature(
+                    projectFlag,
+                    environmentFlags[id],
+                    'identity-overrides',
+                  )
                 }}
                 count={projectFlag.num_identity_overrides}
               />
@@ -305,6 +317,9 @@ class TheComponent extends Component {
                     <Tag className='chip--xs' tag={Constants.archivedTag} />
                   )}
                 </TagValues>
+                {!!isCompact && (
+                  <StaleFlagWarning projectFlag={projectFlag} />
+                )}
               </Row>
               {description && !isCompact && (
                 <div
@@ -312,6 +327,7 @@ class TheComponent extends Component {
                   style={{ lineHeight: '20px', width: width[4] }}
                 >
                   {description}
+                  <StaleFlagWarning projectFlag={projectFlag} />
                 </div>
               )}
             </Flex>